# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School ERP API",
    "version": "2.0.0",
    "description": "School Management System API with RBAC and secure cookie-based authentication",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.yourschool.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token in Authorization header (backward compatibility)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "JWT access token stored in HTTP-only cookie (recommended)"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/images/{key}": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Proxy endpoint to stream images from S3",
        "description": "Proxy endpoint to stream images from S3 without exposing S3 URLs",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The S3 object key (stored in DB as photoKey). Can contain slashes (e.g., tenants/123/students/photo.jpg)"
          }
        ],
        "responses": {
          "200": {
            "description": "Image stream from S3"
          },
          "400": {
            "description": "Invalid image key"
          },
          "404": {
            "description": "Image not found"
          },
          "500": {
            "description": "Failed to retrieve image"
          }
        }
      }
    },
...
  }
}